+++ examples/box_drawing-test.rs
@@ -1,7 +1,7 @@
 /*
     examples/box_drawing-test.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -20,6 +20,7 @@
     IN THE SOFTWARE.
 */
 
+extern crate gettextrs;
 extern crate ncurseswwin;
 extern crate rand;
 extern crate strum;
@@ -27,6 +28,7 @@ extern crate strum_macros;
 
 use std::{time, collections::HashMap};
 
+use gettextrs::*;
 use ncurseswwin::{*, normal::*};
 
 use rand::prelude::*;
@@ -51,7 +53,7 @@ fn main() {
 }
 
 fn main_routine() -> result!(()) {
-    setlocale(LcCategory::All, "")?;
+    setlocale(LocaleCategory::LcAll, "");
 
     // initialize ncurses in a safe way.
     ncursesw_entry(|window| {