+++ examples/menu-test.rs
@@ -1,7 +1,7 @@
 /*
     examples/menu-test.rs
 
-    Copyright (c) 2019 Stephen Whittle  All rights reserved.
+    Copyright (c) 2019, 2020 Stephen Whittle  All rights reserved.
 
     Permission is hereby granted, free of charge, to any person obtaining a copy
     of this software and associated documentation files (the "Software"),
@@ -26,8 +26,10 @@
     indicating that the action has been performed.
 */
 
+extern crate gettextrs;
 extern crate ncurseswwin;
 
+use gettextrs::*;
 use ncurseswwin::{*, menu::*};
 
 macro_rules! result { ($t: ty) => { Result<$t, NCurseswWinError> } }
@@ -42,7 +44,7 @@ fn main() {
 }
 
 fn main_routine() -> result!(()) {
-    setlocale(LcCategory::All, "")?;
+    setlocale(LocaleCategory::LcAll, "");
 
     // initialize ncurses in a safe way.
     ncursesw_entry(|window| {