+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 
 ## [0.5.0] - ????-??-?? [BC]
 - NCurses form module implemented as `form`.
-- `menu_request_by_name()` now returns a `Result<Option<ncursesw::menu::MenuRequest>, ncursesw::NCurseswMenuError>` instead of `Result<bool, ncursesw::NCurseswMenuError>`. [BC]
+- Added NCurses screen functionality.
+- Removed `setlocale()` function. [BC]
+- `menu_request_by_name()` now returns a `Result<Option<menu::MenuRequest>, menu::NCurseswMenuError>` instead of `Result<bool, menu::NCurseswMenuError>`. [BC]
 - `menu::menu_request_name()` now takes a `menu::MenuRequest` instead of a `i32` for the request. [BC]
 - `menu::Menu::set_menu_pad()` now takes a `char` instead of a `i32` for the pad character. [BC]
 - `menu::Menu::{set_item_init, set_item_term, set_menu_init, set_menu_term}` now use `Fn(&Menu)` traits. [BC]
@@ -12,11 +14,12 @@ All breaking changes are marked with [BC] and potentially require API consumer c
 - `menu::PostedMenu::menu_driver()` now returns a `Result<Option<menu::MenuRequest>, menu::NCurseswMenuError>` instead of `Result<Option<i32>, menu::NCurseswMenuError>`. [BC]
 - `BaseCanvas::putwin()`, `IsWindow::putwin()` and `IsPad::putwin()` now take `O: std::os::unix::io::AsRawFd + std::io::Write` instead of `&std::path::Path`. [BC]
 - `NCurseswWindow::getwin()` and `IsPad::genwin()` now take `I: std::os::unix::io::AsRawFd + std::io::Read` instead of `&std::path::Path`. [BC]
+- Moved `is_subwin()` from `Scrollable` trait to `BaseCanvas` trait. [BC]
 
 ## [0.4.0] - 2019-12-09 [BC]
 - Coordinate system's used within the crate (i.e. `Origin`, `Size` etc.) have been changed from using axis basic types of `i32` to `u16`. [BC]
 - NCurses menu module implemented as `menu`.
-- Moved `dupwin()` and `getwin()` from the `IsWindow` trait to `NCursesWindow` trait [BC].
+- Moved `dupwin()` and `getwin()` from the `IsWindow` trait to `NCursesWindow` trait. [BC]
 
 ## [0.3.1] - 2019-11-07
 - NCurses `mouse` functionality added.